renderer/requirements.txt
aio-pika>=9.5.4
manim-voiceover[gtts]
numpy
